Implement task return values and fix unsoundness in task spawning
The unsoundness was introduced in d268104f, when `scope` started taking
a `Future` instead of a closure, since to support this, we started
passing the `Scope` by value rather than by reference, which allowed
tasks to spawn their own subtasks. We weren't able to find a way to
allow this in a manner that is sound, so we've removed this capability
instead.